cleanup and move around some utils
