+++ packages/cloudflare/src/cli/deploy/deploy.ts
@@ -1,0 +1,4 @@
+import { BuildOptions } from "@opennextjs/aws/build/helper.js";
+import { OpenNextConfig } from "@opennextjs/aws/types/open-next.js";
+
+export async function deploy(options: BuildOptions, config: OpenNextConfig) {}
